requirements.txt
langchain-ollama
streamlit
pdfplumber
python-dotenv
pymongo
